--- global/utils/auth/logoutUtil.tsx
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2022 The Ontario Institute for Cancer Research. All rights reserved
- *
- * This program and the accompanying materials are made available under the terms of
- * the GNU Affero General Public License v3.0. You should have received a copy of the
- * GNU Affero General Public License along with this program.
- *  If not, see <http://www.gnu.org/licenses/>.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
- * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
- * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
- * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
- * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-import deleteTokens from './deleteTokens';
-import Router from 'next/router';
-
-const logoutUtil = ({
-  handleTokenState = () => {},
-  isManual = false,
-  isPublic = false,
-}: {
-  handleTokenState: ((jwt: string) => void | PromiseLike<void>) | undefined;
-  isManual: boolean;
-  isPublic?: boolean;
-}) => {
-  console.log('logoutUtil');
-
-  deleteTokens().then(() => {
-    handleTokenState('');
-  });
-
-  if (isManual || !isPublic) {
-    Router.push({
-      pathname: '/',
-      ...(isManual
-        ? {}
-        : {
-            query: { session_expired: true },
-          }),
-    });
-  }
-};
-
-export default logoutUtil;
